Move useLoadableQuery defer tests to own file
