+++ src/react/hooks/__tests__/useLoadableQuery.test.tsx
@@ -1531,163 +1531,6 @@ it("works with startTransition to change variables", async () => {
   });
 });
 
-it('does not suspend deferred queries with data in the cache and using a "cache-and-network" fetch policy', async () => {
-  interface Data {
-    greeting: {
-      __typename: string;
-      message: string;
-      recipient: { name: string; __typename: string };
-    };
-  }
-
-  const query: TypedDocumentNode<Data, Record<string, never>> = gql`
-    query {
-      greeting {
-        message
-        ... @defer {
-          recipient {
-            name
-          }
-        }
-      }
-    }
-  `;
-
-  const link = new MockSubscriptionLink();
-  const cache = new InMemoryCache();
-  cache.writeQuery({
-    query,
-    data: {
-      greeting: {
-        __typename: "Greeting",
-        message: "Hello cached",
-        recipient: { __typename: "Person", name: "Cached Alice" },
-      },
-    },
-  });
-  const client = new ApolloClient({
-    cache,
-    link,
-    incrementalHandler: new Defer20220824Handler(),
-  });
-
-  using _disabledAct = disableActEnvironment();
-  const renderStream = createDefaultProfiler<Data>();
-  const { SuspenseFallback, ReadQueryHook } =
-    createDefaultProfiledComponents(renderStream);
-
-  function App() {
-    useTrackRenders();
-    const [loadQuery, queryRef] = useLoadableQuery(query, {
-      fetchPolicy: "cache-and-network",
-    });
-    return (
-      <div>
-        <button onClick={() => loadQuery()}>Load todo</button>
-        <Suspense fallback={<SuspenseFallback />}>
-          {queryRef && <ReadQueryHook queryRef={queryRef} />}
-        </Suspense>
-      </div>
-    );
-  }
-
-  const { user } = await renderWithClient(
-    <App />,
-    {
-      client,
-    },
-    renderStream
-  );
-
-  // initial render
-  await renderStream.takeRender();
-
-  await user.click(screen.getByText("Load todo"));
-
-  {
-    const { snapshot, renderedComponents } = await renderStream.takeRender();
-
-    expect(renderedComponents).toStrictEqual([App, ReadQueryHook]);
-
-    expect(snapshot.result).toStrictEqualTyped({
-      data: {
-        greeting: {
-          __typename: "Greeting",
-          message: "Hello cached",
-          recipient: { __typename: "Person", name: "Cached Alice" },
-        },
-      },
-      dataState: "complete",
-      error: undefined,
-      networkStatus: NetworkStatus.loading,
-    });
-  }
-
-  link.simulateResult({
-    result: {
-      data: {
-        greeting: { __typename: "Greeting", message: "Hello world" },
-      },
-      hasNext: true,
-    },
-  });
-
-  {
-    const { snapshot, renderedComponents } = await renderStream.takeRender();
-
-    expect(renderedComponents).toStrictEqual([ReadQueryHook]);
-    expect(snapshot.result).toStrictEqualTyped({
-      data: {
-        greeting: {
-          __typename: "Greeting",
-          message: "Hello world",
-          recipient: { __typename: "Person", name: "Cached Alice" },
-        },
-      },
-      dataState: "streaming",
-      error: undefined,
-      networkStatus: NetworkStatus.streaming,
-    });
-  }
-
-  link.simulateResult(
-    {
-      result: {
-        incremental: [
-          {
-            data: {
-              recipient: { name: "Alice", __typename: "Person" },
-              __typename: "Greeting",
-            },
-            path: ["greeting"],
-          },
-        ],
-        hasNext: false,
-      },
-    },
-    true
-  );
-
-  {
-    const { snapshot, renderedComponents } = await renderStream.takeRender();
-
-    expect(renderedComponents).toStrictEqual([ReadQueryHook]);
-    expect(snapshot.result).toStrictEqualTyped({
-      data: {
-        greeting: {
-          __typename: "Greeting",
-          message: "Hello world",
-          recipient: { __typename: "Person", name: "Alice" },
-        },
-      },
-      dataState: "complete",
-      error: undefined,
-      networkStatus: NetworkStatus.ready,
-    });
-  }
-
-  await expect(renderStream).not.toRerender();
-});
 
 it("reacts to cache updates", async () => {
   const { query, mocks } = useSimpleQueryCase();
@@ -4553,173 +4396,6 @@ it('suspends and does not use partial data when changing variables and using a "
   }
 });
 
-it('does not suspend deferred queries with partial data in the cache and using a "cache-first" fetch policy with `returnPartialData`', async () => {
-  interface QueryData {
-    greeting: {
-      __typename: string;
-      message?: string;
-      recipient?: {
-        __typename: string;
-        name: string;
-      };
-    };
-  }
-
-  const query: TypedDocumentNode<QueryData, Record<string, never>> = gql`
-    query {
-      greeting {
-        message
-        ... on Greeting @defer {
-          recipient {
-            name
-          }
-        }
-      }
-    }
-  `;
-
-  const link = new MockSubscriptionLink();
-  const cache = new InMemoryCache();
-
-  {
-    // We are intentionally writing partial data to the cache. Supress console
-    // warnings to avoid unnecessary noise in the test.
-    using _consoleSpy = spyOnConsole("error");
-
-    cache.writeQuery({
-      query,
-      data: {
-        greeting: {
-          __typename: "Greeting",
-          recipient: { __typename: "Person", name: "Cached Alice" },
-        },
-      },
-    });
-  }
-
-  const client = new ApolloClient({
-    link,
-    cache,
-    incrementalHandler: new Defer20220824Handler(),
-  });
-
-  using _disabledAct = disableActEnvironment();
-  const renderStream = createDefaultProfiler<DeepPartial<QueryData>>();
-  const { SuspenseFallback, ReadQueryHook } =
-    createDefaultProfiledComponents(renderStream);
-
-  function App() {
-    useTrackRenders();
-    const [loadTodo, queryRef] = useLoadableQuery(query, {
-      fetchPolicy: "cache-first",
-      returnPartialData: true,
-    });
-
-    return (
-      <div>
-        <button onClick={() => loadTodo()}>Load todo</button>
-        <Suspense fallback={<SuspenseFallback />}>
-          {queryRef && <ReadQueryHook queryRef={queryRef} />}
-        </Suspense>
-      </div>
-    );
-  }
-
-  const { user } = await renderWithClient(
-    <App />,
-    {
-      client,
-    },
-    renderStream
-  );
-
-  // initial render
-  await renderStream.takeRender();
-
-  await user.click(screen.getByText("Load todo"));
-
-  {
-    const { snapshot, renderedComponents } = await renderStream.takeRender();
-
-    expect(renderedComponents).toStrictEqual([App, ReadQueryHook]);
-    expect(snapshot.result).toStrictEqualTyped({
-      data: {
-        greeting: {
-          __typename: "Greeting",
-          recipient: { __typename: "Person", name: "Cached Alice" },
-        },
-      },
-      dataState: "partial",
-      error: undefined,
-      networkStatus: NetworkStatus.loading,
-    });
-  }
-
-  link.simulateResult({
-    result: {
-      data: {
-        greeting: { message: "Hello world", __typename: "Greeting" },
-      },
-      hasNext: true,
-    },
-  });
-
-  {
-    const { snapshot, renderedComponents } = await renderStream.takeRender();
-
-    expect(renderedComponents).toStrictEqual([ReadQueryHook]);
-    expect(snapshot.result).toStrictEqualTyped({
-      data: {
-        greeting: {
-          __typename: "Greeting",
-          message: "Hello world",
-          recipient: { __typename: "Person", name: "Cached Alice" },
-        },
-      },
-      dataState: "streaming",
-      error: undefined,
-      networkStatus: NetworkStatus.streaming,
-    });
-  }
-
-  link.simulateResult(
-    {
-      result: {
-        incremental: [
-          {
-            data: {
-              __typename: "Greeting",
-              recipient: { name: "Alice", __typename: "Person" },
-            },
-            path: ["greeting"],
-          },
-        ],
-        hasNext: false,
-      },
-    },
-    true
-  );
-
-  {
-    const { snapshot, renderedComponents } = await renderStream.takeRender();
-
-    expect(renderedComponents).toStrictEqual([ReadQueryHook]);
-    expect(snapshot.result).toStrictEqualTyped({
-      data: {
-        greeting: {
-          __typename: "Greeting",
-          message: "Hello world",
-          recipient: { __typename: "Person", name: "Alice" },
-        },
-      },
-      dataState: "complete",
-      error: undefined,
-      networkStatus: NetworkStatus.ready,
-    });
-  }
-
-  await expect(renderStream).not.toRerender();
-});
 
 it("throws when calling loadQuery on first render", async () => {
   // We don't provide this functionality with React 19 anymore since it requires internals access
